### 2022-03-17 #4247 @editorialbot
[PRE REVIEW]: PyNM: a Lightweight Python implementation of Normative Modeling
**Submitting author:** @harveyaa (<a href="http://orcid.org/0000-0002-9940-8799">Annabelle Harvey</a>)
**Repository:** https://github.com/ppsp-team/PyNM
**Branch with paper.md** (empty if default branch): joss_submission
**Version:** 1.0.0b1
