### 2022-03-17 #4247 @kthyng
Hi @harveyaa and thanks for your submission! I saw your note that earlier versions of this software have been used in papers. This is fine, as long as the scholarly effort separate from those papers is large enough to justify another publication. Could you go into detail about what the other papers have been about (do they describe the software in detail or simply use the software?) and how much development has occurred on the software since the last paper? Thanks.
