### 2022-03-18 #4247 @deep-introspection
Hi @kthyng! Only rudimentary implementations of LOESS/Centiles methods were used in the past papers. Not only the PyNM library covers other more advanced methods (GP and GAMLSS especially), but even the LOESS/Centiles implementations have been totally refactored for creating a real API. Overall, PyNM also contains many additions: a command-line interface for wide accessibility, automatic dataset splitting and cross-validation, five models from various back-ends in a unified interface that covers a broad range of common use cases, solutions for very large datasets and heteroscedastic data, integrated plotting and evaluation functions to quickly check the validity of the model fit and results, and comprehensive and interactive tutorials.
