### 2022-03-19 #4247 @kthyng
@harveyaa, @deep-introspection thanks for that information. I will ping the greater editorial board with that information available so they can help decide if this is in scope for JOSS; this process will take 1-2 weeks. Thank you!
