### 2022-04-08 #4247 @dfm
🎉 Thanks @smkia for agreeing to review this submission!

In our discussions, @smkia noted a potential [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy): a [publication from 2019](https://www.sciencedirect.com/science/article/abs/pii/S0149763419303197) co-authored with @deep-introspection, a co-author of this submission. My understanding is that @smkia and @deep-introspection did not have direct contact related to that project, and neigher were the primary author, so I'd be happy to waive this conflict if the submitting authors (@harveyaa & @deep-introspection) agree. Please let me know here or [via email](mailto:dfm@dfm.io) if you have any concerns with this resolution.
