### 2022-04-15 #4321 @editorialbot
[REVIEW]: PyNM: a Lightweight Python implementation of Normative Modeling
**Submitting author:** @harveyaa (<a href="http://orcid.org/0000-0002-9940-8799">Annabelle Harvey</a>)
**Repository:** https://github.com/ppsp-team/PyNM
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.1
