### 2022-04-15 #4321 @dfm
@harveyaa — Can you take a look at fixing the DOI issue [mentioned above](https://github.com/openjournals/joss-reviews/issues/4321#issuecomment-1100391043) by the bot?
