+++ requirements.txt
@@ -1,11 +1,11 @@
-matplotlib
-seaborn
-tqdm
-numpy >= 1.21.0
+gpytorch >= 1.4.0
+matplotlib == 3.3.4
+numpy == 1.19.5
 pandas >= 1.1.5
+rpy2 >= 3.1.0
 scikit_learn >= 0.24.1
 scipy >= 1.5.3
+seaborn
 statsmodels >= 0.13.2
-rpy2 >= 3.1.0
 torch >= 1.8.0
-gpytorch >= 1.4.0
+tqdm
