### 2022-04-27 #4321 @harveyaa
Hi,

I made a couple changes to the `joss_submission` branch to fix an import error a colleague caught - hopefully this doesn't confuse the review process too much - should I update pypi with the change or leave it in github for now?

Thank you,

Annabelle
