### 2022-05-03 #4321 @dfm
@harveyaa — Sorry for the delayed response! We generally suggest minting a new release at the end of the review process, but you're also welcome to release new versions throughout the process and it shouldn't cause any problems.
