### 2022-06-03 48b7b74
Update Statsmodels dependency
