### 2022-06-03 6b65078
Update setup.py for subfolder models
