### 2022-06-04 #4321 @smkia
Checking the functionality of the package, I am trying to run the first tutorial at https://github.com/ppsp-team/PyNM/blob/master/tutorials/1-getting_started.ipynb in Google Colab (see https://colab.research.google.com/drive/1LNEIT2T2omOr3iKa4qqUPz9VlwaxeuNG?usp=sharing). I receive the following error:

`ModuleNotFoundError                       Traceback (most recent call last)
[<ipython-input-2-1a8dca4f882a>](https://localhost:8080/#) in <module>()
      1 import pandas as pd
      2 import numpy as np
----> 3 from pynm.pynm import PyNM
      4 import matplotlib.pyplot as plt
      5 import seaborn as sns

[/usr/local/lib/python3.7/dist-packages/pynm/pynm.py](https://localhost:8080/#) in <module>()
     36
     37 from pynm.util import *
---> 38 from pynm.models.loess import *
     39 from pynm.models.centiles import *
     40

ModuleNotFoundError: No module named 'pynm.models'
'

I suspect to version difference between the package and code in the tutorial. Would please hint at that.
