### 2022-06-04 #4321 @dfm
@smkia: Can you open this issue over on https://github.com/ppsp-team/PyNM/issues, referencing this thread instead? That tends to be a better workflow than putting issues in this thread. Thanks!
