### 2022-06-05 a73f689
Update setup.py for dependencies
