### 2022-07-05 ppsp-team/PyNM#23 @saigerutherford
documentation for testing
There is no documentation on how to test, and this is a requirement on the [JOSS checklist](https://github.com/openjournals/joss-reviews/issues/4321). When running `test/test_pynm.py` I also get an error message that the pytest package is not installed in my virtual environment, so this package should be added to `requirements.txt` and some documentation for how to run tests should be added to the repository.
