### 2022-07-05 ppsp-team/PyNM#24 @saigerutherford
update some references and consider adding a few
Regarding the [JOSS checklist (references)](https://github.com/openjournals/joss-reviews/issues/4321) the following should be addressed:

1. The PCNToolkit repository has a Zenodo DOI and the citation should be: Andre Marquand, Saige Rutherford, Seyed Mostafa Kia, Thomas Wolfers, Charlotte Fraza, Richard Dinga, & Mariam Zabihi. (2021). PCNToolkit (0.20). Zenodo. https://doi.org/10.5281/zenodo.5207839
2. The Normative Modeling Framework for Computational Psychiatry preprint has been published and should be cited as: Rutherford, S., Kia, S.M., Wolfers, T. et al. The normative modeling framework for computational psychiatry. Nat Protoc 17, 1711–1734 (2022). https://doi.org/10.1038/s41596-022-00696-5
3. When mentioning GAMLSS implementation, this work should also be cited: [Dinga et al. 2021](https://doi.org/10.1101/2021.06.14.448106)
4. Minor comment, but when mentioning the PCNToolkit development and applications, consider citing: [Rutherford et al. 2022](https://elifesciences.org/articles/72904) & [Kia et al. 2021](https://doi.org/10.1101/2021.05.28.446120)
