### 2022-07-05 ppsp-team/PyNM#25 @saigerutherford
false claim regarding PCNtoolkit usability for beginners
Re: [JOSS review](https://github.com/openjournals/joss-reviews/issues/4321)

In the paper, there is a sentence about the PCNToolkit stating: "While it includes features that make it an obvious choice for advanced users in many cases, is not as approachable to beginners and does not implement several key models."

From my perspective, this is an unfair critique. Instead of a vague/general statement that your package is more approachable to beginners (which you do not make clear how it is more approachable to beginners), I think it is more productive to focus on specific examples of how your package is different from the PCNToolkit, like implementing different algorithms (i.e., GAMLSS).

I do have to admit that I am biased here, because in my Ph.D. I have invested tons of time to make the PCNToolkit accessible to beginners with an in-depth [protocol paper](https://t.co/pm4wVBpuXV), lots of [tutorials](https://pcntoolkit.readthedocs.io/en/latest/pages/BLR_normativemodel_protocol.html) that are easily accessible in a web browser (no local python install required), a [forum](https://gitter.im/predictive-clinical-neuroscience/community) for communication/asking questions, etc.

So, while I think that your package does have certain benefits that should be highlighted, I do not agree that it is more accessible for beginners than the PCNToolkit.
