+++ pynm/util.py
@@ -59,7 +59,7 @@ def SMSE(y_true,y_pred):
     """
     return (np.mean((y_true - y_pred)**2))**0.5/np.std(y_true)
 
-def MSLL(y_true,y_pred,sigma):
+def MSLL(y_true,y_pred,sigma,y_train_mean,y_train_sigma):
     """Calculates Mean Standardized Log Loss (MSLL).
 
     Parameters
@@ -70,14 +70,27 @@ def MSLL(y_true,y_pred,sigma):
         Predicted values for response variable
     sigma: array
         Standard deviation of predictive distribution.
+    y_train_mean: float
+        Mean of training data.
+    y_train_sigma: float
+        Standard deviation of training data.
+    
     
     Returns
     -------
     float
         MSLL value for inputs.
     """
-    SLL = (0.5 * np.log(2 * np.pi * sigma**2) +
-                    (y_true - y_pred)**2 / (2 * sigma**2) -
-                    (y_true - np.mean(y_true))**2 /
-                    (2 * np.std(y_true)))
+    #SLL = (0.5 * np.log(2 * np.pi * sigma**2) +
+    #                (y_true - y_pred)**2 / (2 * sigma**2) -
+    #                (y_true - np.mean(y_true))**2 /
+    #                (2 * np.std(y_true)))
+
+    # Negative log probability under the model
+    model = (0.5*np.log(2*np.pi*sigma**2)) + (y_true - y_pred)**2/(2*sigma**2)
+
+    # Negative log probability under trivial model
+    trivial = (0.5*np.log(2*np.pi*y_train_sigma**2)) + (y_true - y_train_mean)**2/(2*y_train_sigma**2)
+
+    SLL = model - trivial
     return np.mean(SLL)
