### 2022-07-06 ppsp-team/PyNM#26 @saigerutherford
nilearn missing from requirements.txt
Re: [JOSS review](https://github.com/openjournals/joss-reviews/issues/4321) functionality

When running the tutorials, I get an error in tutorials `5-model_selection.ipynb` and `6-downstream_analysis.ipynb` because nilearn is not installed in the virtual environment. Please add nilearn to `requirements.txt`.
