### 2022-07-06 ppsp-team/PyNM#27 @saigerutherford
tutorial 4 error due to R installation
Re: [JOSS review](https://github.com/openjournals/joss-reviews/issues/4321) functionality

I am unable to run the `4-complex_data.ipynb` tutorial because I do not have R installed on my laptop, and there are no clear instructions for this requirement and how to fulfill it. It also reads in the paper as if you have written the GAMLSS R functions in python, which does not seem to be true based on the tutorial errors. Please clarify this in the installation instructions and in the paper.
