### 2022-07-11 808f70c
added pytest to requirements.txt
