+++ paper/paper.md
@@ -54,7 +54,7 @@ The basic idea underpinning Normative Modeling is to fit a model on the controls
 
 In principle, any model that estimates both the mean and variance of the predictive distribution could be used for Normative Modeling. However, in practice we impose more constraints. First and foremost, the assumptions of the model must be met by the data. Second, we want to distinguish between epistemic and aleatoric uncertainty. Epistemic or systematic uncertainty stems from how information about the distribution is collected, whereas aleatoric uncertainty is intrinsic to the distribution and represents the true variation of the population [@xu:2021].
 
-To the author’s knowledge, PCNtoolkit [@pcntoolkit] is the only other available package for Normative Modeling. It implements methods that have been applied in a range of psychiatry and neuroimaging studies including [@kia:2020] and [@fraza:2021], and is accompanied by thorough tutorials and a framework for Normative Modeling in computational psychiatry [@rutherford:2021]. While it includes features that make it an obvious choice for advanced users in many cases, is not as approachable to beginners and does not implement several key models.
+To the author’s knowledge, PCNtoolkit [@pcntoolkit] is the only other available package for Normative Modeling. It implements methods that have been applied in a range of psychiatry and neuroimaging studies including [@kia:2020] and [@fraza:2021], and is accompanied by thorough tutorials and a framework for Normative Modeling in computational psychiatry [@rutherford:2022]. While it includes features that make it an obvious choice for advanced users in many cases, is not as approachable to beginners and does not implement several key models.
 
 PyNM is intended to take users through their first steps in Normative Modeling to using advanced models on complex datasets. Crucially, it manages the dataset and has interactive tutorials – making it quick for new users to try the method either on their own data or on provided simulated data. The tutorials motivate the use of each model and highlight their limitations to help clarify which model is appropriate for what data, and built in plotting and evaluation functions (\autoref{fig:Figure 1}) make it simple to check the validity of the model output. The package includes five models from various backends in a unified interface, including a wrapper for GAMLSS [@rigby:2005] from R that is otherwise not yet available in python, and the selected models cover many settings including big data and heteroskedasticity.
 
