### 2022-08-23 69bc102
fixed paper.bib trailing whitespace
