+++ paper/paper.bib
@@ -98,7 +98,7 @@
     }
 
 @Misc{pcntoolkit,
-  author = {Andre Marquand 
+  author = {Andre Marquand
   and Saige Rutherford
   and Seyed Mostafa Kia
   and Thomas Wolfers
