+++ requirements.txt
@@ -5,7 +5,7 @@ pandas >= 1.1.5
 pytest == 6.2.3
 rpy2 == 3.1.0
 scikit_learn == 0.24.1
-scipy == 1.5.3
+scipy >= 1.5.3
 seaborn == 0.11.1
 statsmodels == 0.12.2
 torch == 1.8.0
