+++ paper/paper.bib
@@ -139,6 +139,28 @@
     isbn="978-3-030-59728-3"
     }
 
+@Article {kia:2021,
+	author = {Kia, Seyed Mostafa
+    and Huijsdens, Hester
+    and Rutherford, Saige
+    and Dinga, Richard
+    and Wolfers, Thomas
+    and Mennes, Maarten
+    and Andreassen, Ole A.
+    and Westlye, Lars T.
+    and Beckmann, Christian F.
+    and Marquand, Andre F.},
+	title = {Federated Multi-Site Normative Modeling using Hierarchical Bayesian Regression},
+	elocation-id = {2021.05.28.446120},
+	year = {2021},
+	doi = {10.1101/2021.05.28.446120},
+	publisher = {Cold Spring Harbor Laboratory},
+	abstract = {Clinical neuroimaging data availability has grown substantially in the last decade, providing the potential for studying heterogeneity in clinical cohorts on a previously unprecedented scale. Normative modeling is an emerging statistical tool for dissecting heterogeneity in complex brain disorders. However, its application remains technically challenging due to medical data privacy issues and difficulties in dealing with nuisance variation, such as the variability in the image acquisition process. Here, we introduce a federated probabilistic framework using hierarchical Bayesian regression (HBR) for multi-site normative modeling. The proposed method completes the life-cycle of normative modeling by providing the possibilities to learn, update, and adapt the model parameters on decentralized neuroimaging data. Our experimental results confirm the superiority of HBR in deriving more accurate normative ranges on large multi-site neuroimaging datasets compared to the current standard methods. In addition, our approach provides the possibility to recalibrate and reuse the learned model on local datasets and even on datasets with very small sample sizes. The proposed federated framework closes the technical loop for applying normative modeling across multiple sites in a decentralized manner. This will facilitate applications of normative modeling as a medical tool for screening the biological deviations in individuals affected by complex illnesses such as mental disorders.Competing Interest StatementOle A. Andreassen is a consultant to HealthLytix and received a speaker honorarium from Lundbeck. Christian F. Beckmann is a shareholder and director of SBG Neuro.},
+	URL = {https://www.biorxiv.org/content/early/2021/05/30/2021.05.28.446120},
+	eprint = {https://www.biorxiv.org/content/early/2021/05/30/2021.05.28.446120.full.pdf},
+	journal = {bioRxiv}
+    }
+
 @Article{fraza:2021,
     title = {Warped Bayesian linear regression for normative modelling of big data},
     journal = {NeuroImage},
