+++ paper/paper.bib
@@ -161,6 +161,24 @@
 	journal = {bioRxiv}
     }
 
+@Article {dinga:2021,
+	author = {Dinga, Richard
+    and Fraza, Charlotte J.
+    and Bayer, Johanna M.M.
+    and Kia, Seyed Mostafa
+    and Beckmann, Christian F.
+    and Marquand, Andre F.},
+	title = {Normative modeling of neuroimaging data using generalized additive models of location scale and shape},
+	elocation-id = {2021.06.14.448106},
+	year = {2021},
+	doi = {10.1101/2021.06.14.448106},
+	publisher = {Cold Spring Harbor Laboratory},
+	abstract = {Normative modeling aims to quantify the degree to which an individual{\textquoteright}s brain deviates from a reference sample with respect to one or more variables, which can be used as a potential biomarker of a healthy brain and as a tool to study heterogeneity of psychiatric disorders. The application of normative models is hindered by methodological challenges and lacks standards for the usage and evaluation of normative models. In this paper, we present generalized additive models for location scale and shape (GAMLSS) for normative modeling of neuroimaging data, a flexible modeling framework that can model heteroskedasticity, non-linear effects of variables, and hierarchical structure of the data. It can model non-Gaussian distributions, and it allows for an automatic model order selection, thus improving the accuracy of normative models while mitigating problems of overfitting. Furthermore, we describe measures and diagnostic tools suitable for evaluating normative models and step-by-step examples of normative modeling, including fitting several candidate models, selecting the best models, and transferring them to new scan sites.Competing Interest StatementThe authors have declared no competing interest.},
+	URL = {https://www.biorxiv.org/content/early/2021/06/14/2021.06.14.448106},
+	eprint = {https://www.biorxiv.org/content/early/2021/06/14/2021.06.14.448106.full.pdf},
+	journal = {bioRxiv}
+}
+
 @Article{fraza:2021,
     title = {Warped Bayesian linear regression for normative modelling of big data},
     journal = {NeuroImage},
