+++ requirements.txt
@@ -1,6 +1,6 @@
 gpytorch >= 1.4.0
-matplotlib == 3.3.4
-numpy == 1.22.0
+matplotlib >= 3.3.4
+numpy >= 1.19.5
 pandas >= 1.1.5
 pytest == 6.2.3
 rpy2 == 3.1.0
