+++ .binder/install.R
@@ -1,0 +1,4 @@
+# Install R packages at build time install.packages() syntax
+install.packages("gamlss")
+install.packages("gamlss.dist")
+install.packages("gamlss.data")
