+++ .binder/environment.yml
@@ -1,0 +1,7 @@
+channels:
+  - conda-forge
+
+dependencies:
+  - nilearn
+  - pip:
+    - pynm
