+++ .binder/environment.yml
@@ -2,6 +2,5 @@ channels:
   - conda-forge
 
 dependencies:
-  - nilearn
   - pip:
     - pynm
