+++ .binder/environment.yml
@@ -1,3 +1,5 @@
+name: python 3.9
+
 channels:
   - conda-forge
 
