+++ requirements.txt
@@ -7,6 +7,6 @@ rpy2 == 3.1.0
 scikit_learn == 0.24.1
 scipy >= 1.5.3
 seaborn == 0.11.1
-statsmodels == 0.12.2
+statsmodels == 0.13.2
 torch == 1.8.0
 tqdm == 4.59.0
