### 2022-09-05 f25fc3d
updated install_requires
