+++ setup.py
@@ -25,16 +25,16 @@ setup(
     },
    install_requires=[
         'gpytorch >= 1.4.0',
-        'matplotlib >= 3.3.4',
-        'numpy >= 1.21.0',
+        'matplotlib',
+        'numpy',
         'pandas >= 1.1.5',
-        'pytest >= 6.2.3'
+        'pytest',
         'rpy2 >= 3.1.0',
         'scikit_learn >= 0.24.1',
         'scipy >= 1.5.3',
-        'seaborn >= 0.11.1',
+        'seaborn',
         'statsmodels >= 0.13.2',
         'torch >= 1.8.0',
-        'tqdm >= 4.59.0',
+        'tqdm',
    ],
 )
