### 2022-09-07 #4321 @dfm
@harveyaa — It looks like @saigerutherford and @smkia have a couple of open issues on the https://github.com/ppsp-team/PyNM repo. Can you update on your status tackling these issues? Thanks!!
