### 2022-09-12 40c1f63
removed pytest from install requires
