### 2022-09-12 ppsp-team/PyNM#23 @harveyaa
Hi @saigerutherford, I added pytest to `requirements.txt` and a section to the `README.md` entitled "How to run tests" for documentation on testing the code.
