### 2022-09-12 ppsp-team/PyNM#26 @harveyaa
Hi @saigerutherford, nilearn is only used in that tutorial as a part of example usage of pynm and isn't required for the package itself. In order to keep the package lightweight I added notebook cells to the relevant tutorials with a note about nilearn installation and the code to do so on the spot rather than putting nilearn in the `requirements.txt`.
