### 2022-09-12 ppsp-team/PyNM#27 @harveyaa
Hi @saigerutherford, re R installation instructions:
I added a link to the documentation for installing R in the `README.md` and state that it is a necessary prerequisite for using the GAMLSS model. This seems fair to me since the installation is system specific and outside of the python package universe it seems out of the scope of the pynm documentation.

re GAMLSS implementation reference in the paper:
The paper reads "...including a wrapper for GAMLSS [@rigby:2005] from R that is otherwise not yet available in python...", which I feel makes it clear the GAMLSS functions are in R and have not been implemented in python (by me or anyone to my knowledge).
