### 2022-09-12 2542995
added rutherford 2022 to paper refs
