### 2022-09-12 ppsp-team/PyNM#24 @harveyaa
Hi @saigerutherford, I made all the updates/additions you mentioned.
