### 2022-09-12 ppsp-team/PyNM#25 @harveyaa
Hi @saigerutherford, I updated the paper to read:

"To the author’s knowledge, PCNtoolkit [@pcntoolkit] is the only other available package for Normative Modeling. It implements methods that have been applied in a range of psychiatry and neuroimaging studies including [@kia:2020], [@kia:2021], [@rutherford:2022a] and [@fraza:2021], and is accompanied by thorough [tutorials](https://pcntoolkit.readthedocs.io/en/latest/pages/BLR_normativemodel_protocol.html), a [forum](https://gitter.im/predictive-clinical-neuroscience/community), and a framework for Normative Modeling in computational psychiatry [@rutherford:2022b]. While PCNtoolkit offers more advanced functionality, PyNM emphasizes being lightweight and easy to use, and implements different models than PCNtookit including the GAMLSS which is a powerful option for Normative Modeling [@dinga:2021]."

Which removes the critique in favour of highlighting the accompanying paper, forum and tutorials (that make PCNtoolkit accessible to beginners) and mentioning pynm's differences from PCNtoolkit.
