### 2022-09-22 791a276
fixed statsmodels warning unknown kwarg weight
