+++ test/test_pynm.py
@@ -613,7 +613,7 @@ class TestCV:
         assert 'GAMLSS_pred' in m.data.columns
     
     def test_cv_3_gamlss(self):
-        data = generate_data(sample_size=4, n_sites=2, randseed=3)
+        data = generate_data(sample_size=5, n_sites=2, randseed=3)
         m = pynm.PyNM(data,'score','group',['age','c(sex)','c(site)'])
         m.gamlss_normative_model(mu='score ~ cs(age)',sigma='~ age + site',tau='~ c(sex)',cv_folds=3)
         assert 'GAMLSS_pred' in m.data.columns
