+++ pynm/pynm.py
@@ -22,7 +22,6 @@
 
 import pandas as pd
 import numpy as np
-import warnings
 import matplotlib.pyplot as plt
 import seaborn as sns
 import warnings
@@ -816,7 +815,7 @@ class PyNM:
                     # For MSLL
                     y_train_mean = np.mean(self.data[self.score].values[train_index])
                     y_train_sigma = np.std(self.data[self.score].values[train_index])
-
+                    
                     gamlss.fit(X_train)
 
                     cv_mu_pred = gamlss.predict(X_test,what='mu')
