+++ .binder/environment.yml
@@ -1,4 +1,4 @@
-name: python 3.7
+name: python 3.9
 
 channels:
   - conda-forge
