+++ .binder/environment.yml
@@ -4,5 +4,7 @@ channels:
   - conda-forge
 
 dependencies:
+  - r-base=4.2
+  - r-tidyverse
   - pip:
     - pynm
