### 2022-10-20 0190e8b
Delete install.R
