+++ .binder/runtime.txt
@@ -1,3 +1,0 @@
-# R version
-#r-4.0-2022-08-12
-r-4.2.1
