### 2022-11-14 1a71242
adjusted dim of input to MSLL
