+++ requirements.txt
@@ -6,7 +6,7 @@ pytest == 6.2.3
 rpy2 == 3.5.4
 scikit_learn == 1.1.2
 scipy >= 1.5.3
-seaborn == 0.12.0
+seaborn >= 0.12.0
 statsmodels == 0.13.2
 torch == 1.12.1
-tqdm == 4.59.0
+tqdm >= 4.59.0
