### 2022-11-30 #4321 @smkia
Dear @dfm, the package passes all my tests and I have no further concerns regarding the functionality of the implemented modules. I think this is a useful resource for new users of NM and, therefore, I support this work being published.
