+++ paper/paper.bib
@@ -98,7 +98,7 @@
     }
 
 @Misc{pcntoolkit,
-  author = {Andre Marquand
+  author = {Andre F. Marquand
   and Saige Rutherford
   and Seyed Mostafa Kia
   and Thomas Wolfers
@@ -274,17 +274,18 @@
     url={https://doi.org/10.1038/s41596-022-00696-5}
     }
 
-@Article{rigby:2005,
-    title = {Generalized additive models for location, scale and
-      shape,(with discussion)},
-    journal = {Applied Statistics},
-    volume = {54},
-    part = {3},
-    pages = {507-554},
-    year = {2005},
-    author = {R. A. Rigby 
-    and D. M. Stasinopoulos},
-  }
+@article{rigby:2005,
+author = {Rigby, R. A. and Stasinopoulos, D. M.},
+title = {Generalized additive models for location, scale and shape},
+journal = {Journal of the Royal Statistical Society: Series C (Applied Statistics)},
+volume = {54},
+number = {3},
+pages = {507-554},
+doi = {https://doi.org/10.1111/j.1467-9876.2005.00510.x},
+url = {https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/j.1467-9876.2005.00510.x},
+eprint = {https://rss.onlinelibrary.wiley.com/doi/pdf/10.1111/j.1467-9876.2005.00510.x},
+year = {2005}
+}
 
 @Article{lefebvre:2018,
     author={Lefebvre, Aline 
