+++ requirements.txt
@@ -2,11 +2,11 @@ gpytorch >= 1.4.0
 matplotlib >= 3.3.4
 numpy >= 1.19.5
 pandas >= 1.1.5
-pytest == 6.2.3
-rpy2 == 3.5.4
-scikit_learn == 1.1.2
+pytest >= 6.2.3
+rpy2 >= 3.5.4
+scikit_learn >= 1.1.2
 scipy >= 1.5.3
 seaborn >= 0.12.0
-statsmodels == 0.13.2
-torch == 1.12.1
+statsmodels >= 0.13.2
+torch >= 1.12.1
 tqdm >= 4.59.0
