### 2022-12-04 #4321 @deep-introspection
Thanks a lot, @smkia, @saigerutherford, and @dfm! Both reviews and editorial handling have really enhanced the library and associated manuscript. As requested, I have now integrated your modifications and here is the Zenodo DOI for the v1.0.1 release: [10.5281/zenodo.7396721](https://zenodo.org/record/7396721).
