### 2022-12-06 #4321 @dfm
@deep-introspection, @harveyaa — Thanks! Can you update the metadata on the Zenodo entry so that the title and author list of the archive are the same as the paper? Then please take a look at the generated PDF 👇 to make sure you're happy with it.
