+++ paper/paper.md
@@ -19,11 +19,11 @@ authors:
     affiliation: "1, 2"
   - name: Dumas, Guillaume
     orcid: 0000-0002-2253-1844
-    affiliation: "1, 3"
+    affiliation: "2, 3"
 affiliations:
- - name: Centre de Recherche du CHU Sainte-Justine, Université de Montréal, QC, Canada
-   index: 1
  - name: Centre de Recherche de l’Institut Universitaire de Gériatrie de Montréal, Université de Montréal, QC, Canada
+   index: 1
+ - name: Centre de Recherche du CHU Sainte-Justine, Université de Montréal, QC, Canada
    index: 2
  - name: Mila - Quebec AI Institute, Université de Montréal, QC, Canada
    index: 3
@@ -36,7 +36,7 @@ bibliography: paper.bib
 
 The majority of studies in neuroimaging and psychiatry are focussed on case-control analysis [@marquand:2019]. However, case-control relies on well-defined groups which is more the exception than the rule in biology. Psychiatric conditions are diagnosed based on symptoms alone, which makes for heterogeneity at the biological level [@marquand:2016]. Relying on mean differences obscures this heterogeneity and the resulting loss of information can produce unreliable results or misleading conclusions [@loth:2021].
 
-Normative Modeling is an emerging alternative to case-control analyses, that seeks to parse heterogeneity by looking at how individuals deviate from the normal trajectory. Analogous to normative growth charts, normative models map the mean and variance of a trait for a given population against a set of explanatory variables (usually including age). Statistical inferences at the level of the individual participant can then be obtained with respect to the normative range [@marquand:2019]. This framework can detect patterns of abnormality that might not be consistent across the population, and recasts disease as an extreme of the normal range.
+Normative Modeling is an emerging alternative to case-control analyses that seeks to parse heterogeneity by looking at how individuals deviate from the normal trajectory. Analogous to normative growth charts, normative models map the mean and variance of a trait for a given population against a set of explanatory variables (usually including age). Statistical inferences at the level of the individual participant can then be obtained with respect to the normative range [@marquand:2019]. This framework can detect patterns of abnormality that might not be consistent across the population, and recasts disease as an extreme deviation from the normal range rather than a separate group.
 
 PyNM is a lightweight python implementation of Normative Modeling making it approachable and easy to adopt. The package provides:
 
@@ -52,9 +52,9 @@ PyNM is a lightweight python implementation of Normative Modeling making it appr
 
 The basic idea underpinning Normative Modeling is to fit a model on the controls (or a subset of them) of a dataset, and then apply it to the rest of the participants. The difference between the model’s prediction and the ground truth for the unseen participants relative to the variance around the prediction quantifies their deviation from the normal. While simple in concept, implementing Normative Modeling requires some care in managing the dataset and choosing an appropriate model.
 
-In principle, any model that estimates both the mean and variance of the predictive distribution could be used for Normative Modeling. However, in practice, we impose more constraints. First and foremost, the assumptions of the model must be met by the data. Second, we want to distinguish between epistemic and aleatoric uncertainty. Epistemic or systematic uncertainty stems from how information about the distribution is collected, whereas aleatoric uncertainty is intrinsic to the distribution and represents the true variation of the population [@xu:2021].
[…]
