### 2022-12-08 #4321 @dfm
@deep-introspection — Thanks! We're getting close.

- I fixed a small syntax error in the new reviewer acknowledgement: https://github.com/ppsp-team/PyNM/pull/33
- The title of the zenodo archive is still showing up as "ppsp-team/PyNM: v1.0.1", could you update that to "PyNM: a Lightweight Python implementation of Normative Modeling"? It's not the end of the world if you can't, but we prefer to have an exact match on the title for bookkeeping purposes. Thanks!
