+++ paper/paper.bib
@@ -281,7 +281,7 @@ journal = {Journal of the Royal Statistical Society: Series C (Applied Statistic
 volume = {54},
 number = {3},
 pages = {507-554},
-doi = {https://doi.org/10.1111/j.1467-9876.2005.00510.x},
+doi = {10.1111/j.1467-9876.2005.00510.x},
 url = {https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/j.1467-9876.2005.00510.x},
 eprint = {https://rss.onlinelibrary.wiley.com/doi/pdf/10.1111/j.1467-9876.2005.00510.x},
 year = {2005}
