### 2022-12-08 #4321 @dfm
Thanks @deep-introspection!! With apologies I've opened one (last! I hope) PR fixing a syntax error that I introduced in an earlier PR. Very sorry about that!
