### 2022-12-08 #4321 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
